refactor: 검색 단위 컴포넌트들 components 폴더로 이동
