+++ src/pages/Search/components/RelatedSearches/RelatedSearches.tsx
