+++ src/pages/Search/components/SearchBar/RollingBanner.tsx
